ZEN-2473 Added check for remote sync
Without -f, publish will fail if the working tree is not in sync with its branch origin (i.e. git push or git pull is required)

Note that a perfectly reasonable use of -f is to revert the web
content without having to revert the master branch. E.g. you publish a
content update and then discover a serious content mistake. You can
locally check out a prior version of master, then publish with -f to
back off to the corresponding web content. Then fix the problems in
master and republish.