ZEN-2473 Address git bug
git diff-index has a bug wherein files that are touched (mod time
changed but no content or file mode change) are reported as changed by
diff-index, until a subsequent diff is performed, which appears to
reset things. diff-index shows a bogus "destination" SHA1 value of all
zeros. This was causing the post-build git check to fail because it
appeared that the working tree had been made dirty by the build.