ZEN-2474 Propagate path changes in cygwin
Ruby installation and bundler in cygwin updates the windows system path,
but that change is not automatically picked up by invoking scripts. So
now the various install scripts other than install-jekyll write updated
paths, if necessary, to stdout. Invoking scripts are expected to change
their PATH settings before proceeding.

The top-level script advises the user if PATH needs to be updated, and
prints a command that can be copy/pasted to do it.