ZEN-2471 Added -qa option to publish script
You can now say "bin/publish -qa [repo]" to publish to a repo other than
docs. In that case, policy checks are disabled (as if you had said -f).

The repo must already exist, and it must have a master branch. Easiest
way to do that is create it in github and check the "install a README
file" option.

If a gh-pages for that repo already exists it will be removed. Likewise
for a "dummy" branch.

Then the master branch will be cloned in a temp directory, then contents
will be replaced with the current contents of the working tree form
which publish is invoked, in a new branch named dummy. That content will
be committed and pushed to origin/dummy, and then a subtree push will be
used to create a new remote gh-pages branch containing the site.