ZEN-2500 Fix baseurl problem
The site.baseurl variable was serving two purposes, which are normally
not incompatible, but in the qa site they were. First, it was being used
by the local test server (the one run by jekyll serve) as a prefix of
each requested URL to be chopped of before trying to resolve the path in
the site tree. Second, it was being prepended to all the declared
permalinks for internal pages.

The two uses are now split. The site.baseurl variable (defined in
_config.yml) is still '/docs' so local server will work correctly. But
now site.data.baseurl is prepended to permalinks. The latter is a simple
quoted string in _data/baseurl.yml, so it's easy to replace. Normal
value is "/docs" which works in local test server and in the advertised
public site. But when setting up to push to a qa server, the name of the
qa repo is written into this file instead. (That's simpler and less
error-prone than trying to alter the value within the _config.yml file).